requirements.txt
Flask
openai
pandas
python-dotenv
openpyxl
